+++ test/spec/page.spec.js
@@ -8,16 +8,16 @@ describe( 'Pages mode', () => {
             const form = loadForm( 'pages.xml' );
             form.init();
 
-            expect( form.pages.$current[ 0 ] ).toEqual( form.view.html.querySelector( '.question' ) );
+            expect( form.pages.current ).toEqual( form.view.html.querySelector( '.question' ) );
         } );
 
         it( 'loads to the proper first page if the form contains only a repeat and nothing outside it', () => {
             const form = loadForm( 'repeat-only-pages.xml' );
             form.init();
 
-            expect( form.pages.$current.length ).toEqual( 1 );
-            expect( form.pages.$current.hasClass( 'current' ) ).toBe( true );
-            expect( form.pages.$current[ 0 ] ).toEqual( form.view.html.querySelector( '.or-repeat' ) );
+            expect( form.pages.current ).not.toEqual( null );
+            expect( form.pages.current.classList.contains( 'current' ) ).toBe( true );
+            expect( form.pages.current ).toEqual( form.view.html.querySelector( '.or-repeat' ) );
         } );
 
     } );